Youtube_video_summarizer/requirements.txt
yt_dlp==2024.8.6
librosa==0.10.2.post1
langchain==0.2.15
langchain_community==0.2.15
pydub==0.25.1
transformers==4.44.2
torch==2.4.0
python-dotenv==1.0.1
streamlit==1.38.0
langchain_openai==0.1.23